requirements.txt
flask==2.3.3
Flask-RESTful==0.3.10
python-dotenv==1.0.0
requests
coverage==5.3
gunicorn==20.1.0